Project2/server/requirements.txt
yfinance>=0.1.63
numpy>=1.21.0
pandas>=1.3.3
Flask==2.0.2
kafka-python==2.0.2